refactor: remove shared barrels for services and stores
