+++ app/frontend/src/features/history/components/GenerationHistoryController.vue
@@ -24,7 +24,7 @@ import {
   type DimensionFilterOption,
 } from '@/features/history';
 import { PERSISTENCE_KEYS, useAsyncLifecycleTask, usePersistence } from '@/composables/shared';
-import { useBackendRefresh } from '@/services';
+import { useBackendRefresh } from '@/services/system/backendRefresh';
 import { useBackendBase } from '@/utils/backend';
 import { formatHistoryDate } from '@/utils/format';
 import type { GenerationHistoryResult } from '@/types';
