--- app/frontend/src/services/index.ts
@@ -1,3 +1,0 @@
-export * from './apiClient';
-export * from './backendClient';
-export * from './system';
