--- app/frontend/src/services/system/index.ts
@@ -1,4 +1,0 @@
-export * from './systemService';
-export * from './backendEnvironmentEventBus';
-export * from './backendEnvironmentBus';
-export * from './backendRefresh';
