--- app/frontend/src/stores/index.ts
@@ -1,4 +1,0 @@
-export * from './settings';
-export * from './app';
-export * from './adminMetrics';
-
